chore: properly separates server / client exports
