+++ packages/payload/src/exports/config.ts
@@ -1,7 +1,7 @@
 export { buildConfig } from '../config/build'
 export * from '../config/types'
 
-export { type FieldTypes, fieldTypes } from '../admin/components/forms/field-types'
+export { type FieldTypes } from '../admin/components/forms/field-types'
 export { defaults } from '../config/defaults'
 export { sanitizeConfig } from '../config/sanitize'
 export { baseBlockFields } from '../fields/baseFields/baseBlockFields'
