+++ packages/payload/src/exports/utilities.ts
@@ -1,10 +1,11 @@
+export { withMergedProps } from '../admin/components/utilities/WithMergedProps'
 export { extractTranslations } from '../translations/extractTranslations'
-export { i18nInit } from '../translations/init'
 
+export { i18nInit } from '../translations/init'
 export { combineMerge } from '../utilities/combineMerge'
 export { configToJSONSchema, entityToJSONSchema } from '../utilities/configToJSONSchema'
-export { createArrayFromCommaDelineated } from '../utilities/createArrayFromCommaDelineated'
 
+export { createArrayFromCommaDelineated } from '../utilities/createArrayFromCommaDelineated'
 export { deepCopyObject } from '../utilities/deepCopyObject'
 export { deepMerge } from '../utilities/deepMerge'
 export { default as flattenTopLevelFields } from '../utilities/flattenTopLevelFields'
