+++ packages/richtext-lexical/src/field/features/Link/drawer/index.tsx
@@ -2,7 +2,7 @@ import { Drawer } from 'payload/components/elements'
 import { Form } from 'payload/components/forms'
 import { RenderFields } from 'payload/components/forms'
 import { FormSubmit } from 'payload/components/forms'
-import { fieldTypes } from 'payload/config'
+import { fieldTypes } from 'payload/components/forms'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
